Fix issue after merge (data type change)

+++ code/widgets/org.eclipse.scout.widgets.old.client/src/main/java/org/eclipse/scout/widgets/old/client/ui/forms/BrowserFieldSandboxForm.java
@@ -111,7 +111,7 @@ public class BrowserFieldSandboxForm extends AbstractForm {
         }
 
         @Override
-        protected int getConfiguredLabelPosition() {
+        protected byte getConfiguredLabelPosition() {
           return LABEL_POSITION_TOP;
         }
